% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbnardl.R
\name{coef.fbnardl}
\alias{coef.fbnardl}
\title{Extract coefficients from fbnardl object}
\usage{
\method{coef}{fbnardl}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"fbnardl"}}

\item{...}{Additional arguments (ignored)}
}
\description{
Extract coefficients from fbnardl object
}
\value{
A named numeric vector of model coefficients.
}
