% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds_test.R
\name{compute_bounds_test}
\alias{compute_bounds_test}
\title{Compute Bounds or Bootstrap Cointegration Test}
\usage{
compute_bounds_test(
  model,
  data,
  depvar,
  decompose,
  control_vars,
  type,
  reps,
  fourier,
  kstar,
  p,
  q,
  r,
  nobs
)
}
\arguments{
\item{model}{Estimated lm object}

\item{data}{Data frame with decomposed variables}

\item{depvar}{Name of dependent variable}

\item{decompose}{Names of decomposed variables}

\item{control_vars}{Names of control variables}

\item{type}{"fnardl" for analytical CV, "fbnardl" for bootstrap}

\item{reps}{Number of bootstrap replications}

\item{fourier}{Logical, include Fourier terms}

\item{kstar}{Optimal Fourier frequency}

\item{p}{Optimal lag for dependent variable}

\item{q}{Optimal lags for decomposed variables}

\item{r}{Optimal lags for control variables}

\item{nobs}{Number of observations}
}
\value{
List with test statistics and critical values
}
\description{
Compute Bounds or Bootstrap Cointegration Test
}
\keyword{internal}
