% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{compute_dynamic_multipliers}
\alias{compute_dynamic_multipliers}
\title{Compute Dynamic Multipliers}
\usage{
compute_dynamic_multipliers(model, depvar, decompose, p, horizon)
}
\arguments{
\item{model}{Estimated lm object}

\item{depvar}{Name of dependent variable}

\item{decompose}{Names of decomposed variables}

\item{p}{Lag order for dependent variable}

\item{horizon}{Number of periods for multiplier computation}
}
\value{
List with dynamic multiplier paths
}
\description{
Following Shin, Yu and Greenwood-Nimmo (2014), compute cumulative
dynamic multipliers for positive and negative shocks
}
\keyword{internal}
