% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compute_multipliers}
\alias{compute_multipliers}
\title{Compute Short-Run and Long-Run Multipliers}
\usage{
compute_multipliers(model, depvar, decompose, control_vars, q, r)
}
\arguments{
\item{model}{Estimated lm object}

\item{depvar}{Name of dependent variable}

\item{decompose}{Names of decomposed variables}

\item{control_vars}{Names of control variables}

\item{q}{Optimal lags for decomposed variables}

\item{r}{Optimal lags for control variables}
}
\value{
List with SR and LR multipliers for each variable
}
\description{
Compute Short-Run and Long-Run Multipliers
}
\keyword{internal}
