% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{estimate_final_model}
\alias{estimate_final_model}
\title{Estimate Final Model}
\usage{
estimate_final_model(
  data,
  depvar,
  decompose,
  control_vars,
  p,
  q,
  r,
  kstar,
  nobs,
  fourier
)
}
\arguments{
\item{data}{Data frame with decomposed variables}

\item{depvar}{Name of dependent variable}

\item{decompose}{Names of decomposed variables}

\item{control_vars}{Names of control variables}

\item{p}{Optimal lag for dependent variable}

\item{q}{Optimal lags for decomposed variables}

\item{r}{Optimal lags for control variables}

\item{kstar}{Optimal Fourier frequency}

\item{nobs}{Number of observations}

\item{fourier}{Logical, include Fourier terms}
}
\value{
lm object with estimated model
}
\description{
Estimate Final Model
}
\keyword{internal}
