% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbnardl.R
\name{fitted.fbnardl}
\alias{fitted.fbnardl}
\title{Extract fitted values from fbnardl object}
\usage{
\method{fitted}{fbnardl}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"fbnardl"}}

\item{...}{Additional arguments (ignored)}
}
\description{
Extract fitted values from fbnardl object
}
\value{
A numeric vector of fitted values from the model.
}
