% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{jb_test}
\alias{jb_test}
\title{Jarque-Bera Normality Test}
\usage{
jb_test(resid)
}
\arguments{
\item{resid}{Residual vector}
}
\value{
List with JB statistic, df, and p-value
}
\description{
Jarque-Bera Normality Test
}
\keyword{internal}
