% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbnardl.R
\name{nobs.fbnardl}
\alias{nobs.fbnardl}
\title{Extract number of observations from fbnardl object}
\usage{
\method{nobs}{fbnardl}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"fbnardl"}}

\item{...}{Additional arguments (ignored)}
}
\description{
Extract number of observations from fbnardl object
}
\value{
An integer giving the number of observations used in the model fit.
}
