% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbnardl.R
\name{plot.fbnardl}
\alias{plot.fbnardl}
\title{Plot method for fbnardl objects}
\usage{
\method{plot}{fbnardl}(x, type = c("all", "multipliers", "kstar", "residuals"), ...)
}
\arguments{
\item{x}{An object of class \code{"fbnardl"}}

\item{type}{Character string specifying the plot type. One of
\code{"multipliers"}, \code{"kstar"}, \code{"residuals"}, or \code{"all"}.}

\item{...}{Additional arguments passed to plotting functions}
}
\description{
Plot method for fbnardl objects
}
\value{
No return value, called for side effects (generates diagnostic plots).
}
