% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_dynamic_multipliers}
\alias{plot_dynamic_multipliers}
\title{Plot Dynamic Multipliers}
\usage{
plot_dynamic_multipliers(dyn_mult, decomposed_vars)
}
\arguments{
\item{dyn_mult}{List of dynamic multipliers from compute_dynamic_multipliers}

\item{decomposed_vars}{Names of decomposed variables}
}
\description{
Plot Dynamic Multipliers
}
\value{
No return value, called for side effects.
}
\keyword{internal}
