% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_residual_diagnostics}
\alias{plot_residual_diagnostics}
\title{Plot Residual Diagnostics}
\usage{
plot_residual_diagnostics(x)
}
\arguments{
\item{x}{fbnardl object}
}
\description{
Plot Residual Diagnostics
}
\value{
No return value, called for side effects.
}
\keyword{internal}
