% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{print_asymmetry_tests}
\alias{print_asymmetry_tests}
\title{Print Asymmetry Tests}
\usage{
print_asymmetry_tests(tests, decomposed_vars)
}
\arguments{
\item{tests}{List of asymmetry test results}

\item{decomposed_vars}{Names of decomposed variables}
}
\description{
Print Asymmetry Tests
}
\value{
No return value, called for side effects.
}
\keyword{internal}
