% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{print_bounds_test}
\alias{print_bounds_test}
\title{Print Bounds Test Results}
\usage{
print_bounds_test(test, type)
}
\arguments{
\item{test}{List with bounds test results}

\item{type}{"fnardl" or "fbnardl"}
}
\description{
Print Bounds Test Results
}
\value{
No return value, called for side effects.
}
\keyword{internal}
