% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{print_multipliers}
\alias{print_multipliers}
\title{Print Multipliers Table}
\usage{
print_multipliers(multipliers, decomposed_vars, control_vars)
}
\arguments{
\item{multipliers}{List of multipliers}

\item{decomposed_vars}{Names of decomposed variables}

\item{control_vars}{Names of control variables}
}
\description{
Print Multipliers Table
}
\value{
No return value, called for side effects.
}
\keyword{internal}
