% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbnardl.R
\name{residuals.fbnardl}
\alias{residuals.fbnardl}
\title{Extract residuals from fbnardl object}
\usage{
\method{residuals}{fbnardl}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"fbnardl"}}

\item{...}{Additional arguments (ignored)}
}
\description{
Extract residuals from fbnardl object
}
\value{
A numeric vector of residuals from the model.
}
