% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbnardl.R
\name{vcov.fbnardl}
\alias{vcov.fbnardl}
\title{Extract variance-covariance matrix from fbnardl object}
\usage{
\method{vcov}{fbnardl}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"fbnardl"}}

\item{...}{Additional arguments (ignored)}
}
\description{
Extract variance-covariance matrix from fbnardl object
}
\value{
A variance-covariance matrix of the model coefficients.
}
