% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds_test.R
\name{bootstrap_cointegration_test}
\alias{bootstrap_cointegration_test}
\title{Bootstrap Cointegration Test}
\usage{
bootstrap_cointegration_test(
  model,
  data,
  depvar,
  decompose,
  control_vars,
  fourier,
  kstar,
  p,
  q,
  r,
  nobs,
  reps,
  level_vars,
  indep_level_vars
)
}
\arguments{
\item{model}{Estimated lm object}

\item{data}{Data frame}

\item{depvar}{Name of dependent variable}

\item{decompose}{Names of decomposed variables}

\item{control_vars}{Names of control variables}

\item{fourier}{Logical}

\item{kstar}{Fourier frequency}

\item{p, q, r}{Lag orders}

\item{nobs}{Number of observations}

\item{reps}{Number of bootstrap replications}

\item{level_vars}{Names of all level variables}

\item{indep_level_vars}{Names of independent level variables}
}
\value{
List with bootstrap critical values and p-values
}
\description{
Implements the bootstrap procedure of Bertelli, Vacca & Zoia (2022)
and McNown, Sam & Goh (2018)
}
\keyword{internal}
