% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{print_diagnostics}
\alias{print_diagnostics}
\title{Print Diagnostic Tests}
\usage{
print_diagnostics(diag)
}
\arguments{
\item{diag}{List of diagnostic test results}
}
\description{
Print Diagnostic Tests
}
\keyword{internal}
